Use git in the subproject instead of the root project
If a subproject has it's own git repository
use this one instead of the root project.

For example if repo is used the subprojects
itself can also be repositories and in some
configurations they should be handled differently
than the root project that is why the
git repository from the subproject is now used.