Allow to configure the version file encoding
Without this setting, the JVM default encoding is used.
This can easily break files if they contain non-ASCII characters.
A properties file for example by definintion is ISO-8859-1 encoded.
Java 18+ use UTF-8 as default encoding if nothing else is configured.
If you now manipulate the version file using UTF-8, it gets crippled
if it contains non-ASCII characters.
With this new setting the encoding can be configured, with the
appropriate default setting according to the default for the file.